Extract public functions into utils
