+++ hosts.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"net"
 	"os"
-	"regexp"
 	"strings"
 	"sync"
 	"time"
@@ -239,16 +238,4 @@ func (f *FileHosts) Refresh() {
 
 func (f *FileHosts) clear() {
 	f.hosts = make(map[string]string)
-}
-
-func isDomain(domain string) bool {
-	if isIP(domain) {
-		return false
-	}
-	match, _ := regexp.MatchString(`^([a-zA-Z0-9\*]([a-zA-Z0-9\-]{0,61}[a-zA-Z0-9])?\.)+[a-zA-Z]{2,6}$`, domain)
-	return match
-}
-
-func isIP(ip string) bool {
-	return (net.ParseIP(ip) != nil)
 }
